update App: fix logger, fix test
