--- src/main/java/com/tinkoff/edu/app/LoanCalcController.java
@@ -1,14 +1,0 @@
-package com.tinkoff.edu.app;
-
-/**
- *Controller
- */
-public class LoanCalcController {
-    /**
-     * Validates and logs request
-     */
-    public static int createRequest() {
-        LoanCalcLogger.log();
-        return LoanCalcService.createRequest();
-    }
-}
