--- src/main/java/com/tinkoff/edu/app/LoanCalcLogger.java
@@ -1,10 +1,0 @@
-package com.tinkoff.edu.app;
-
-public class LoanCalcLogger {
-    /**
-     *our logs
-     */
-    public static void log() {
-        System.out.println("ATTENTION! \nERROR! \nSOS!");
-    }
-}
