--- src/main/java/com/tinkoff/edu/app/LoanCalcRepository.java
@@ -1,12 +1,0 @@
-package com.tinkoff.edu.app;
-
-public class LoanCalcRepository {
-    private static int requestId;
-    /**
-     * TODO persists request
-     * @return RequestId
-     */
-    public static int save() {
-        return ++requestId;
-    }
-}
