--- src/main/java/com/tinkoff/edu/app/LoanCalcService.java
@@ -1,10 +1,0 @@
-package com.tinkoff.edu.app;
-
-public class LoanCalcService {
-    /**
-     * Loan calculation
-     */
-    public static int createRequest() {
-      return LoanCalcRepository.save();
-    }
-}
